+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
+## 1.6.2 (27.05.2019) 
+ + Add `volumes` and `automount` parameters to `LKDev\HetznerCloud\Models\Servers\Servers` - `createInDatacenter` and `createInLocation`
+ + Add `created` property to `LKDev\HetznerCloud\Models\Servers\Server`
+
 ## 1.6.1 (27.05.2019)
- + Add `automount` and `format` Parameters to `LKDev\HetznerCloud\Models\Volumes\Volumes` - `create`
+ + Add `automount` and `format` parameters to `LKDev\HetznerCloud\Models\Volumes\Volumes` - `create`
  + Add `created` property to `LKDev\HetznerCloud\Models\FloatingIps\FloatingIp`
  + Improve test coverage
  + Add `root_passwort` to response of `LKDev\HetznerCloud\Models\Servers\Server` - `rebuildFromImage` (https://github.com/LKDevelopment/hetzner-cloud-php-sdk/issues/17)